remove instructions val and dependencies
