+++ contrib/src/instructions/bitwise.rs
@@ -1,11 +1,10 @@
 use symbolic_stack_machines_core::{
+    environment::EnvExtension,
     instructions::{
-        AbstractExecRecord, AbstractInstruction, ConcreteAbstractExecRecord,
-        InstructionResult,
+        AbstractExecRecord, AbstractInstruction, ConcreteAbstractExecRecord, InstructionResult,
     },
     memory::Mem,
     stack::{Stack, StackOpRecord, StackRecord},
-    environment::EnvExtension
 };
 
 // The MachineEq trait is necessary because the type of equality used in the machine
