+++ core/src/lib.rs
@@ -1,7 +1,7 @@
 pub mod constraint;
+pub mod environment;
 pub mod instructions;
 pub mod machine;
 pub mod memory;
 pub mod stack;
 pub mod value;
-pub mod environment;
