+++ core/src/machine/mod.rs
@@ -7,12 +7,9 @@ use std::rc::Rc;
 
 use crate::instructions::*;
 use crate::memory::ReadOnlyMem;
-use crate::{
-    memory::{memory_models::MemIntToInt, RWMem},
-    stack::*,
-};
+use crate::{memory::RWMem, stack::*};
 use error::MachineError;
-use z3::ast::{Bool, Int};
+use z3::ast::Bool;
 use z3::{Context, Model, SatResult, Solver};
 
 pub type MachineResult<T> = Result<T, MachineError>;
@@ -48,8 +45,7 @@ where
     I: VMInstruction<'a, Mem = Mem, ValStack = MachineStack>,
     StackVal: Into<MemIdx> + Into<MemVal>,
 {
-    pub fn new(stack: MachineStack, mem_init: Mem::InitArgs) -> Self {
-        let mem = Mem::init(mem_init);
+    pub fn new(stack: MachineStack, mem: Mem) -> Self {
         Self {
             mem,
             stack,
@@ -232,30 +228,5 @@ where
         }
 
         stack.peek(0)
-    }
-}
-
-// Implement machine initialization for a specific memory model
-impl<'a, MachineStack, I>
-    BaseMachine<'a, MemIntToInt<'a>, MachineStack, I, Int<'a>, Int<'a>, Int<'a>>
-where
-    MachineStack: Stack<StackVal = Int<'a>>,
-    I: VMInstruction<'a, Mem = MemIntToInt<'a>, ValStack = MachineStack>,
-{
-    // For symbolic memory
-    pub fn new_with_ctx(stack: MachineStack, mem_init_args: Rc<&'a Context>) -> Self {
-        let mem = MemIntToInt::init(mem_init_args.clone());
-        let ctx = SymbolicContext {
-            constraints: vec![],
-            ctx: mem_init_args.clone(),
-        };
-
-        Self {
-            mem,
-            stack,
-            pgm: vec![],
-            pc: 0,
-            context: Some(ctx),
-        }
     }
 }
