+++ core/src/value/mod.rs
@@ -1,10 +1,10 @@
 mod r#abstract;
 mod ground;
-mod val;
 mod inner;
 mod traits;
-pub use traits::*;
-pub use r#abstract::*;
+mod val;
 pub use ground::*;
-pub use val::*;
 pub use inner::*;
+pub use r#abstract::*;
+pub use traits::*;
+pub use val::*;
